backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
elasticsearch==8.15.1
httpx==0.27.2
python-dotenv==1.0.1
redis==5.0.8
aiofiles==24.1.0
requests==2.32.3
aiohttp==3.9.5
